+++ packages/graphql-codegen-compiler/src/index.ts
@@ -9,4 +9,3 @@ export {
   FlattenOperation
 } from './types';
 export { ALLOWED_CUSTOM_TEMPLATE_EXT } from './generate-multiple-files';
-export { getGeneratorConfig } from 'graphql-codegen-generators';