+++ packages/scripts/handlebars-templates-scripts/cli/codegen-handlebars-templates-scripts.js
@@ -1,0 +1,12 @@
+#!/usr/bin/env node
+
+const program = require('commander');
+const pack = require('../package.json');
+
+program.version(pack.version);
+
+program
+  .command('init', 'Initialize a new template boilerplate in the current directory')
+  .command('build', 'Executes build using Webpack and Handlebars');
+
+program.parse(process.argv);