+++ packages/scripts/handlebars-templates-scripts/webpack.config.js
@@ -1,17 +1,18 @@
 const nodeExternals = require('webpack-node-externals');
 
 module.exports = {
-  context: __dirname + "/src/",
-  entry: "./index.ts",
+  context: '/src/',
+  entry: './index.ts',
   output: {
-    path: __dirname + "/dist/",
-    filename: "index.js",
-    libraryTarget: "commonjs"
+    path: '/dist/',
+    filename: 'index.js',
+    libraryTarget: 'commonjs'
   },
   mode: 'production',
   target: 'node',
   externals: [nodeExternals()],
   resolve: {
+    mainFields: ['browser', 'main', 'module'],
     extensions: ['.ts', '.tsx', '.js', '.jsx']
   },
   devtool: 'source-map',
@@ -26,5 +27,5 @@ module.exports = {
         use: 'raw-loader'
       }
     ]
-  },
+  }
 };